Small library/compilation of 4x4 matrix maths functions
Originally written in Java by me. Contains an
executor function that will take an array of arrays
that contains a string that matches the respective
matrix functions and their numerical arguments
and execute the functions with the arguments in turn.
Does allocate new temporary matrix which is wasteful.
There is also a perspective function.